+++ R/Rules-validity.R
@@ -82,10 +82,6 @@ v_next_best_ncrm_loss <- function(object) {
       "lower bound of unacceptable has to be >= than upper bound of overdose"
     )
   }
-  v$check(
-    test_probability(object@max_overdose_prob, bounds_closed = FALSE),
-    "max_overdose_prob must be a probability value from (0, 1) interval"
-  )
   if (is_unacceptable_valid) {
     losses_len <- ifelse(all(object@unacceptable == c(1, 1)), 3L, 4L)
     v$check(
@@ -320,19 +316,16 @@ v_increments_relative_dlt <- function(object) {
 }
 
 #' @describeIn v_increments validates that the [`IncrementsNumDoseLevels`] object
-#'   contains valid `maxLevels` and `basisLevel` option.
-v_increments_numdoselevels <- function(object) {
+#'   contains valid `max_levels` and `basis_level` option.
+v_increments_num_dose_levels <- function(object) {
   v <- Validate()
   v$check(
-    is.scalar(object@maxLevels) &&
-      is.integer(object@maxLevels) &&
-      object@maxLevels > 0,
-    "maxLevels must be scalar positive integer"
+    test_int(object@max_levels, lower = .Machine$double.xmin),
+    "max_levels must be scalar positive integer"
   )
   v$check(
-    is.scalar(object@basisLevel) &&
-      object@basisLevel %in% c("last", "max"),
-    "basisLevel must be either 'last' or 'max'"
+    test_string(object@basis_level, pattern = "^last$|^max$"),
+    "basis_level must be either 'last' or 'max'"
   )
   v$result()
 }
@@ -343,21 +336,31 @@ v_increments_hsr_beta <- function(object) {
   v <- Validate()
   v$check(
     test_probability(object@target, bounds_closed = FALSE),
-    "target must be a probability"
+    "target must be a probability value from (0, 1) interval"
   )
   v$check(
     test_probability(object@prob, bounds_closed = FALSE),
-    "prob must be a probability"
+    "prob must be a probability value from (0, 1) interval"
   )
   v$check(
-    is.scalar(object@a) & is.numeric(object@a) && object@a > 0,
+    test_number(object@a, lower = .Machine$double.xmin, finite = TRUE),
     "Beta distribution shape parameter a must be a positive scalar"
   )
   v$check(
-    is.scalar(object@b) & is.numeric(object@b) && object@b > 0,
+    test_number(object@b, lower = .Machine$double.xmin, finite = TRUE),
     "Beta distribution shape parameter b must be a positive scalar"
   )
+  v$result()
+}
 
+#' @describeIn v_increments validates that the [`IncrementsMin`]
+#'  object contains a list with `Increments` objects.
+v_increments_min <- function(object) {
+  v <- Validate()
+  v$check(
+    all(sapply(object@increments_list, test_class, "Increments")),
+    "all elements in increments_list must be of Increments class"
+  )
   v$result()
 }
 
@@ -376,9 +379,91 @@ v_increments_hsr_beta <- function(object) {
 #'   or `TRUE` in case validation passes.
 NULL
 
-#' @describeIn v_stopping validates that the [`StoppingLowestDoseHSRBeta`]
-#'  object contains valid probability target, threshold and shape parameters.
-v_stopping_lowest_dose_hsr_beta <- v_increments_hsr_beta
+#' @describeIn v_stopping validates that the [`StoppingCohortsNearDose`]
+#'  object contains valid `nCohorts` and `percentage` parameters.
+v_stopping_cohorts_near_dose <- function(object) {
+  v <- Validate()
+  v$check(
+    test_int(object@nCohorts, lower = .Machine$double.xmin),
+    "nCohorts must be positive integer scalar"
+  )
+  v$check(
+    test_probability(object@percentage / 100),
+    "percentage must be a number between 0 and 100"
+  )
+  v$result()
+}
+
+#' @describeIn v_stopping validates that the [`StoppingPatientsNearDose`]
+#'  object contains valid `nPatients` and `percentage` parameters.
+v_stopping_patients_near_dose <- function(object) {
+  v <- Validate()
+  v$check(
+    test_int(object@nPatients, lower = .Machine$double.xmin),
+    "nPatients must be positive integer scalar"
+  )
+  v$check(
+    test_probability(object@percentage / 100),
+    "percentage must be a number between 0 and 100"
+  )
+  v$result()
+}
+
+#' @describeIn v_stopping validates that the [`StoppingMinCohorts`]
+#'  object contains valid `nCohorts` parameter.
+v_stopping_min_cohorts <- function(object) {
+  v <- Validate()
+  v$check(
+    test_int(object@nCohorts, lower = .Machine$double.xmin),
+    "nCohorts must be positive integer scalar"
+  )
+  v$result()
+}
+
+#' @describeIn v_stopping validates that the [`StoppingMinPatients`]
+#'  object contains valid `nPatients` parameter.
+v_stopping_min_patients <- function(object) {
+  v <- Validate()
+  v$check(
+    test_int(object@nPatients, lower = .Machine$double.xmin),
+    "nPatients must be positive integer scalar"
+  )
+  v$result()
+}
+
+#' @describeIn v_stopping validates that the [`StoppingTargetProb`]
+#'  object contains valid `target` and `prob` parameters.
+v_stopping_target_prob <- function(object) {
+  v <- Validate()
+  v$check(
+    test_probability_range(object@target),
+    "target has to be a probability range"
+  )
+  v$check(
+    test_probability(object@prob, bounds_closed = FALSE),
+    "prob must be a probability value from (0, 1) interval"
+  )
+  v$result()
+}
+
+#' @describeIn v_stopping validates that the [`StoppingMTDdistribution`]
+#'  object contains valid `target`, `thresh` and `prob` parameters.
+v_stopping_mtd_distribution <- function(object) {
+  v <- Validate()
+  v$check(
+    test_probability(object@target, bounds_closed = FALSE),
+    "target must be a probability value from (0, 1) interval"
+  )
+  v$check(
+    test_probability(object@thresh, bounds_closed = FALSE),
+    "thresh must be a probability value from (0, 1) interval"
+  )
+  v$check(
+    test_probability(object@prob, bounds_closed = FALSE),
+    "prob must be a probability value from (0, 1) interval"
+  )
+  v$result()
+}
 
 #' @describeIn v_stopping validates that the [`StoppingMTDCV`] object
 #'   contains valid probability target and percentage threshold.
@@ -389,8 +474,12 @@ v_stopping_mtd_cv <- function(object) {
     "target must be probability value from (0, 1) interval"
   )
   v$check(
-    test_probability(object@thresh_cv / 100, bounds_closed = FALSE),
+    test_probability(object@thresh_cv / 100, bounds_closed = c(FALSE, TRUE)),
     "thresh_cv must be percentage > 0"
   )
   v$result()
 }
+
+#' @describeIn v_stopping validates that the [`StoppingLowestDoseHSRBeta`]
+#'  object contains valid probability target, threshold and shape parameters.
+v_stopping_lowest_dose_hsr_beta <- v_increments_hsr_beta # nolintr