+++ R/Simulations-methods.R
@@ -1,3 +1,4 @@
+# nolint start
 #####################################################################################
 ## Author: Daniel Sabanes Bove [sabanesd *a*t* roche *.* com],
 ##         Wai Yin Yeung [ w *.* yeung1 *a*t* lancaster *.* ac *.* uk]
@@ -2447,5 +2448,4 @@ setMethod("plot",
               ret
             })
 
- ## ------------------------------------------------------------------------
-
+# nolint end